server/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beartype==0.18.5
casefy==0.1.7
certifi==2024.6.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.5
plum-dispatch==2.2.2
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
pyjnius==1.6.1
pyserde==0.19.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.30
starlette==0.37.2
tabulate==0.9.0
typer==0.12.3
typing-extensions==4.12.2
typing-inspect==0.9.0
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
